% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AKSmooth}
\alias{AKSmooth}
\title{AKSmooth, locally kernel smoothing bisulfite sequencing data}
\usage{
AKSmooth(bsdata, h, kernel = c("Gaussian", "Epanechnikov", "Tricube"))
}
\arguments{
\item{bsdata}{Bisulfite sequencing data containing genomic positions, read coverage and beta methylation values.}

\item{h}{The smoothing window size.}

\item{kernel}{The kernel function used for smoothing.}
}
\value{
The estimated DNA methylation value for a single sample.
}
\description{
Modified nadaraya-watson kernel smoother with different kernel functions for the bisulfite sequencing data,
which performs smoothing on low-coverage methylation data chromosome-wise.
}
\details{
\code{bsdata} is a data.frame contains \code{1.)} genomic location, \code{2.)} the chromosomal position,
\code{3.)} M values, describing the number of methylated read covering a single CpG,
\code{4.)} Cov values (read coverage), describing the total number of reads covering a single CpG,
\code{5.)} meth values, describing the beta estimated methylation values computed by M/Cov.
}
\examples{
## load the bisulfite sequencing data (chromosome 21 for one single sample)
data(bn1chr21)
## define a bandwidth of 30 CpGs
fitChr21gau <- AKSmooth(bn1chr21, 30, "Gaussian")
}
\author{
Junfang Chen, Pavlo Lutsik and Ruslan Akulenko.
}

